Refactored script into python package. Implemented simple test.
